Move removeError to reusable hook
